long after one back up 20170824
new analysis and so on

+++ Imaging_data_processing/dft_reg_dir_2_zy.m
@@ -162,7 +162,7 @@ parfor i = 1:length(datafiles)
         fprintf('the shift in this trial is over 20 pixes do rigistration again using itself as target\n');
         match=abs((shift(1,:)))<20 & abs(shift(2,:))<20;
         FrameIsBadAlign{i} = a;
-        im_tg_temp=mean(aa(:,:,match),3);   % change here for the target frame
+        im_tg_temp=mean(double(aa(:,:,match)),3);   % change here for the target frame
         funct=tic;
         [im_dft_reg,shift] = dft_reg(im_s, im_tg_temp); 
         t=toc(funct);
